Move all internal types from /snap-types to /snap-utils
