+++ packages/controllers/src/services/AbstractExecutionService.ts
@@ -1,6 +1,6 @@
 import { Duplex } from 'stream';
 import ObjectMultiplex from '@metamask/object-multiplex';
-import { ErrorJSON, SnapExecutionData } from '@metamask/snap-types';
+import { ErrorJSON, SnapExecutionData } from '@metamask/snap-utils';
 import { SNAP_STREAM_NAMES } from '@metamask/execution-environments';
 import {
   Duration,
