+++ packages/types/README.md
@@ -1,3 +1,3 @@
 # @metamask/snap-types
 
-Types for MetaMask Snaps packages.
+TypeScript types used by MetaMask Snaps at runtime. Contains both a global declaration for the global snap provider (`wallet`) and importable types for e.g. the `onRpcRequest` handler function.
