+++ packages/types/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@metamask/snap-types",
   "version": "0.19.0",
-  "description": "Shared types for MetaMask Snaps packages.",
+  "description": "TypeScript types used by MetaMask Snaps at runtime.",
   "repository": {
     "type": "git",
     "url": "https://github.com/MetaMask/snaps-skunkworks.git"
