+++ packages/types/src/types.d.ts
@@ -1,24 +1,7 @@
-import { Json } from '@metamask/controllers';
 import { MetaMaskInpageProvider } from '@metamask/providers';
 import { JsonRpcRequest } from '@metamask/types';
 
-/**
- * Command request sent to a worker.
- */
-export type WorkerCommandRequest = {
-  id: string;
-  command: string;
-  data?: string | Record<string, unknown>;
-};
-
-export type SnapData = {
-  snapId: string;
-  sourceCode: string;
-};
-
-export type SnapExecutionData = SnapData & {
-  endowments?: Json;
-};
+export { Json, JsonRpcResponse, JsonRpcRequest } from '@metamask/types';
 
 export type SnapRpcHandler = (args: {
   origin: string;
@@ -28,11 +11,3 @@ export type SnapRpcHandler = (args: {
 export type OnRpcRequestHandler = SnapRpcHandler;
 
 export type SnapProvider = MetaMaskInpageProvider;
-
-export type SnapId = string;
-
-export type ErrorJSON = {
-  message: string;
-  code: number;
-  data?: Json;
-};
